Refactor player service design
