+++ backend/src/main/java/com/example/smashboard/service/PlayerService.java
@@ -1,45 +1,16 @@
 package com.example.smashboard.service;
 
 import com.example.smashboard.model.Player;
-import com.example.smashboard.repository.PlayerRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
-@Service
-public class PlayerService {
-    private final PlayerRepository playerRepository;
-
-    public PlayerService(PlayerRepository playerRepository) {
-        this.playerRepository = playerRepository;
-    }
-
-    public Player savePlayer(Player player) {
-        return playerRepository.save(player);
-    }
-
-    public List<Player> getAllPlayers() {
-        return playerRepository.findAll();
-    }
-
-    public Optional<Player> getPlayerByInternalId(Long internalId) {
-        return playerRepository.findById(internalId);
-    }
-
-    public Optional<Player> getPlayerByDuprId(String duprId) {
-        return playerRepository.findById(duprId);
-    }
-
-    public Optional<Player> getPlayerByName(String name) {
-        return playerRepository.findByName(name);
-    }
-
-    public List<Player> searchPlayersByName(String name) {
-        return playerRepository.findByNameContainingIgnoreCase(name);
-    }
-
-    public void deletePlayer(Long internalId) {
-        playerRepository.deleteById(internalId);
-    }
+public interface PlayerService {
+    Player savePlayer(Player player);
+    List<Player> getAllPlayers();
+    Optional<Player> getPlayerByInternalId(Long internalId);
+    Optional<Player> getPlayerByDuprId(String duprId);
+    Optional<Player> getPlayerByName(String name);
+    List<Player> searchPlayersByName(String name);
+    void deletePlayer(Long internalId);
 }
